Rename smartBanner reducer to xpromo; split state up
Refactor notes:

* Navigtating to branch urls is now done in the click handler, in the same call stack
* reducers/smartBanner has been renamed to reducers/xpromo, and split on functionality
  * reducers/xpromo/interstitials now contains all of the 'normal' interstitial state:
    ** showBanner, logingRequired, scrolledPast, scrolledStart
  * Some state keys from reducers/smartBanner has been removed, specificlally:
    ** haveShownXPromo, xPromoshownUrl. they are no longer being
  * reducers/xpromo/listingClick now contains all of the 'listingClick' specific state:
    ** canListingClick, showingListingClickInterstitial

* lib/smartBannerState has been renamed to lib/xpromoState
