+++ src/lib/xpromoState.js
